refactor(progress): step vocabulary registry + step-display split
Consolidate the progress-step vocabulary into a single registry in
domain/progress. Each StepSpec fully describes one step (id, label,
phase, order, applicable surfaces); STEP_LABELS, STEP_PHASES, and the
ordered step lists (steps_for_surface, docker_steps, STEPS_LOCAL) are
all derived from it. register_step() is the runtime extension point for
future modes (server mode's server-start/health/ramp phases): a new step
flows through labels, phase mapping, and ordering with no renderer or
dispatcher edits. The container-boundary alias (preflight to
container_preflight) moves into the registry as resolve_container_step,
replacing the hard-coded mapping in docker_runner (the only docker_runner
hunk: one import plus one line).

Split the 1322-line cli/_step_display.py god-module along its concerns:
_step_render (shared render/formatting helpers), _experiment_display
(StepDisplay), and _study_display (StudyStepDisplay plus row/image
models). _step_display stays a thin facade so CLI callers are unchanged.
The display classes and helpers move verbatim; display output is
byte-identical and the existing display tests pass unchanged.

Add acceptance tests: a hypothetical step registered at runtime flows
through ordering, labels, and phase mapping, and renders via StepDisplay
with no renderer change.
